+++ src/hooks/aelin/useAelinCreateUpFrontDeal.tsx
@@ -5,7 +5,6 @@ import styled from 'styled-components'
 import { isAddress } from '@ethersproject/address'
 import { BigNumber } from '@ethersproject/bignumber'
 import { BigNumberish } from '@ethersproject/bignumber'
-import { parseUnits } from '@ethersproject/units'
 import { wei } from '@synthetixio/wei'
 
 import usePrevious from '../common/usePrevious'
@@ -15,7 +14,7 @@ import { AddressWhitelistProps } from '@/src/components/pools/whitelist/addresse
 import { NftType } from '@/src/components/pools/whitelist/nft/nftWhiteListReducer'
 import { ChainsValues, getKeyChainByValue } from '@/src/constants/chains'
 import { contracts } from '@/src/constants/contracts'
-import { BASE_DECIMALS, EXCHANGE_DECIMALS, ZERO_ADDRESS } from '@/src/constants/misc'
+import { EXCHANGE_DECIMALS, ZERO_ADDRESS } from '@/src/constants/misc'
 import { Privacy } from '@/src/constants/pool'
 import { Token } from '@/src/constants/token'
 import {
@@ -24,11 +23,11 @@ import {
 } from '@/src/hooks/contracts/useAelinDirectDealCreateTransaction'
 import { GasOptions, useTransactionModal } from '@/src/providers/transactionModalProvider'
 import { useWeb3Connection } from '@/src/providers/web3ConnectionProvider'
-import { parseValuesToCreateUpFrontDeal } from '@/src/utils/aelinPoolUtils'
 import { getDuration, isEmptyDuration, secondsToDhm } from '@/src/utils/date'
 import { formatNumber } from '@/src/utils/formatNumber'
 import { mergeArrayKeyValuePairs } from '@/src/utils/mergeArrayKeyValuePairs'
 import { parseBalanceMap } from '@/src/utils/merkle-tree/parse-balance-map'
+import { parseValuesToCreateUpFrontDeal } from '@/src/utils/parseValuesToCreateUpFrontDeal'
 import validateCreateDirectDeal, { dealErrors } from '@/src/utils/validate/createDirectDeal'
 import { storeJson } from '@/src/utils/web3storage'
